+++ platforms/core-runtime/files/src/test/groovy/org/gradle/internal/file/ModificationTimeFileAccessTimeJournalTest.groovy
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.resource.local
+package org.gradle.internal.file
+
 
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
 import org.junit.Rule
